requirements.txt
msal==1.20.0
pyzmq==24.0.1